fix(go): address M-1 and M-2 review findings
M-1: range-binding.ts re-parse violated architecture contract
- Reuse treeCache from orchestrator instead of re-parsing
- Add getTreeSitterBufferSize for fallback parse (DoD §2.6)
- Thread treeCache through populateRangeBindings hook signature

M-2: namespace-import mirroring had no language gate
- Extract namespace typeBinding mirroring from shared
  imported-return-types.ts into Go-specific hook
- Add mirrorNamespaceTypeBindings to ScopeResolver interface
- Go's "uppercase = exported" convention now lives only in Go code
- Export followChainPostFinalize as shared utility
